add test solidity library

+++ src/contract/tests/contracts/library_example.sol
@@ -2,7 +2,7 @@
 pragma solidity >=0.7.0  <0.9.0;
 
 library MyLibrary {
-    function add(uint a, uint b) public pure returns(uint) {
+    function add(uint256 a, uint256 b) public pure returns(uint256) {
         return a + b;
     }
 }
