+++ src/contract/tests/contracts/use_library_example.sol
@@ -5,9 +5,9 @@ import "./library_example.sol";  //导入库
 
 contract UseLibraryExample {
 
-    using MyLibrary for uint;
+    using MyLibrary for uint256;
 
-    function getSum(uint firstNumber, uint secondNumber) public pure returns(uint) {
+    function getSum(uint256 firstNumber, uint256 secondNumber) public pure returns(uint256) {
         return firstNumber.add(secondNumber); //您对这句可能产生了困惑
     }
